requirements.txt
pyrotgfork
python-dotenv
sqlalchemy
tgcrypto
tzdata
rich
pydub
Pillow